+++ python/examples/01_the_switch_and_the_light/light.py
@@ -2,6 +2,7 @@ import atexit
 import sys
 import time
 import RPi.GPIO as GPIO
+from client import Client
 
 GPIO23 = 23
 
@@ -10,8 +11,8 @@ GPIO.setmode(GPIO.BCM)
 GPIO.setup(GPIO23, GPIO.OUT)
 
 # Connect to Arete control plane
-# TODO
-# sys.stderr.write('Connected to Arete control plane\n')
+client = Client.connect('wss://dashboard.test.cns.dev:443')
+sys.stderr.write('Connected to Arete control plane\n')
 
 # Detect future changes in desired state, and actualize it
 # TODO