+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- [#21](https://github.com/project-arete/sdk/issues/21) (Python) Provide an SDK for Python
 - [#16](https://github.com/project-arete/sdk/issues/16) (Rust) Provide an SDK for Rust
 
 ### Changed